tests/final_demo.py
#!/usr/bin/env python3
"""
Final comprehensive demonstration of the permission system
"""

import requests
import json
import time

BASE_URL = "http://localhost:8000"

class PermissionDemo:
    def __init__(self):
        self.admin_token = None
        self.user_token = None
        
    def get_tokens(self):
        """Get authentication tokens"""
        print("🔑 Getting authentication tokens...")
        
        # Get admin token
        admin_response = requests.post(f"{BASE_URL}/login", json={
            "username": "testadmin",
            "password": "admin123"
        })
        if admin_response.status_code == 200:
            self.admin_token = admin_response.json()["tokens"]["access_token"]
            print("✅ Admin token obtained")
        
        # Get user token
        user_response = requests.post(f"{BASE_URL}/login", json={
            "username": "testuser2", 
            "password": "user123"
        })
        if user_response.status_code == 200:
            self.user_token = user_response.json()["tokens"]["access_token"]
            print("✅ User token obtained")
        
        print()

    def demo_public_access(self):
        """Demonstrate public GET access"""
        print("🌐 DEMONSTRATING PUBLIC ACCESS (No Authentication Required)")
        print("=" * 60)
        
        endpoints = [
            ("/categories", "Get all categories"),
            ("/news", "Get published news"),
            ("/news/latest", "Get latest news"),
            ("/health", "Health check")
        ]
        
        for endpoint, description in endpoints:
            response = requests.get(f"{BASE_URL}{endpoint}")
            status = "✅" if response.status_code == 200 else "❌"
            print(f"{status} {description}: {response.status_code}")
            
            # Show sample data for categories
            if endpoint == "/categories" and response.status_code == 200:
                data = response.json()
                print(f"   📊 Found {len(data)} categories")
        
        print()

    def demo_protected_access_without_auth(self):
        """Demonstrate protected endpoints require authentication"""
        print("🔒 DEMONSTRATING PROTECTED ACCESS (Authentication Required)")
        print("=" * 60)
        
        test_cases = [
            ("POST", "/categories", {"name_english": "Test Category"}, "Create category"),
            ("PUT", "/categories/1", {"name_english": "Updated"}, "Update category"),
            ("DELETE", "/categories/999", None, "Delete category"),
            ("GET", "/categories/all", None, "Admin-only endpoint")
        ]
        
        for method, endpoint, data, description in test_cases:
            if method == "POST":
                response = requests.post(f"{BASE_URL}{endpoint}", json=data)
            elif method == "PUT":
                response = requests.put(f"{BASE_URL}{endpoint}", json=data)
            elif method == "DELETE":
                response = requests.delete(f"{BASE_URL}{endpoint}")
            else:
                response = requests.get(f"{BASE_URL}{endpoint}")
            
            status = "✅" if response.status_code == 403 else "❌"
            print(f"{status} {description} without auth: {response.status_code} (Expected 403)")
        
        print()

    def demo_user_vs_admin_access(self):
        """Demonstrate user vs admin access levels"""
        print("👥 DEMONSTRATING ROLE-BASED ACCESS CONTROL")
        print("=" * 60)
        
        if not self.user_token or not self.admin_token:
            print("❌ Missing required tokens for role demonstration")
            return
        
        # Test user access (should be denied)
        print("👤 Testing Regular User Access:")
        user_headers = {"Authorization": f"Bearer {self.user_token}"}
        
        response = requests.post(f"{BASE_URL}/categories", 
                               json={"name_english": "User Test Category"}, 
                               headers=user_headers)
        status = "✅" if response.status_code == 403 else "❌"
        print(f"{status} User creating category: {response.status_code} (Expected 403)")
        
        response = requests.get(f"{BASE_URL}/categories/all", headers=user_headers)
        status = "✅" if response.status_code == 403 else "❌" 
        print(f"{status} User accessing admin endpoint: {response.status_code} (Expected 403)")
        
        # Test admin access (should work)
        print("\n👑 Testing Admin User Access:")
        admin_headers = {"Authorization": f"Bearer {self.admin_token}"}
        
        # Create category
        category_data = {
            "name_english": "Demo Test Category",
            "name_tibetan": "Demo Category Tibet",
            "description_english": "A test category created during demo"
        }
        response = requests.post(f"{BASE_URL}/categories", json=category_data, headers=admin_headers)
        status = "✅" if response.status_code == 201 else "❌"
        print(f"{status} Admin creating category: {response.status_code} (Expected 201)")
        
        category_id = None
        if response.status_code == 201:
            category_id = response.json().get("id")
            print(f"   📝 Created category with ID: {category_id}")
        
        # Access admin endpoint
        response = requests.get(f"{BASE_URL}/categories/all", headers=admin_headers)
        status = "✅" if response.status_code == 200 else "❌"
        print(f"{status} Admin accessing admin endpoint: {response.status_code} (Expected 200)")
        
        # Update category
        if category_id:
            update_data = {"name_english": "Updated Demo Category"}
            response = requests.put(f"{BASE_URL}/categories/{category_id}", 
                                  json=update_data, headers=admin_headers)
            status = "✅" if response.status_code == 200 else "❌"
            print(f"{status} Admin updating category: {response.status_code} (Expected 200)")
            
            # Delete category
            response = requests.delete(f"{BASE_URL}/categories/{category_id}", headers=admin_headers)
            status = "✅" if response.status_code in [200, 204] else "❌"
            print(f"{status} Admin deleting category: {response.status_code} (Expected 200/204)")
        
        print()

    def demo_dashboard_access(self):
        """Demonstrate dashboard admin-only access"""
        print("📊 DEMONSTRATING DASHBOARD ACCESS CONTROL")
        print("=" * 60)
        
        # Test without authentication
        response = requests.get(f"{BASE_URL}/dashboard/stats")
        status = "✅" if response.status_code == 403 else "❌"
        print(f"{status} Dashboard without auth: {response.status_code} (Expected 403)")
        
        # Test with admin token
        if self.admin_token:
            admin_headers = {"Authorization": f"Bearer {self.admin_token}"}
            response = requests.get(f"{BASE_URL}/dashboard/stats", headers=admin_headers)
            status = "✅" if response.status_code == 200 else "❌"
            print(f"{status} Dashboard with admin auth: {response.status_code} (Expected 200)")
            
            if response.status_code == 200:
                stats = response.json()
                print(f"   📈 Total texts: {stats.get('total_texts', 'N/A')}")
                print(f"   📂 Total categories: {stats.get('total_categories', 'N/A')}")
        
        print()

    def demo_query_filtering(self):
        """Demonstrate query parameter filtering"""
        print("🔍 DEMONSTRATING QUERY PARAMETER FILTERING")
        print("=" * 60)
        
        test_cases = [
            ("/categories?lang=en", "English language filtering"),
            ("/categories?lang=tb", "Tibetan language filtering"),
            ("/news?page=1&limit=3", "Pagination filtering"),
            ("/news/latest?limit=2", "Latest news limiting")
        ]
        
        for endpoint, description in test_cases:
            response = requests.get(f"{BASE_URL}{endpoint}")
            status = "✅" if response.status_code == 200 else "❌"
            print(f"{status} {description}: {response.status_code}")
            
            if response.status_code == 200:
                try:
                    data = response.json()
                    if isinstance(data, list):
                        print(f"   📊 Returned {len(data)} items")
                    elif isinstance(data, dict) and 'results' in data:
                        print(f"   📊 Returned {len(data['results'])} items")
                except:
                    pass
        
        print()

    def run_complete_demo(self):
        """Run the complete permission system demonstration"""
        print("🚀 COMPREHENSIVE PERMISSION SYSTEM DEMONSTRATION")
        print("🏛️  Buddhist Digital Library API")
        print("=" * 80)
        print()
        
        self.get_tokens()
        self.demo_public_access()
        self.demo_protected_access_without_auth()
        self.demo_user_vs_admin_access()
        self.demo_dashboard_access()
        self.demo_query_filtering()
        
        print("✅ DEMONSTRATION COMPLETE!")
        print("=" * 80)
        print()
        print("📋 SUMMARY OF IMPLEMENTED FEATURES:")
        print("   ✅ Unified endpoints (no duplicate /admin/* routes)")
        print("   ✅ GET requests work without authentication")
        print("   ✅ POST/PUT/DELETE require admin authentication")
        print("   ✅ Proper HTTP status codes (403 for unauthorized)")
        print("   ✅ Response filtering based on user role")
        print("   ✅ Query parameter filtering")
        print("   ✅ All existing functionality preserved")
        print("   ✅ Admin middleware implementation")
        print("   ✅ Proper error handling")
        print()

def main():
    demo = PermissionDemo()
    demo.run_complete_demo()

if __name__ == "__main__":
    main()
